Extract storage related review summary to storage directory
Keep the affected-systems note (when present) inside the Storage review
description list item instead of a separate one. It's quite weird to
have a row with Label 'Note'.

Also remove IncompleteStepIndicator from the storage code, this should
be added by the Review page.

Use Flex/FlexItem, not Stack for stacking the IncompleteStepIndicator,
so that the items don't get to span the whole row width.

Lastly docs generation had to be adjusted as it errored out with:

file:///home/kkoukiou/repos/anaconda-webui/docs/generate-docs.js:393
            const matchingPage = pages.find(page => page.id === step.id);
