+++ src/components/review/ReviewConfiguration.jsx
@@ -12,7 +12,6 @@ import { DescriptionList } from "@patternfly/react-core/dist/esm/components/Desc
 import { Label } from "@patternfly/react-core/dist/esm/components/Label/index.js";
 import { useWizardContext, useWizardFooter } from "@patternfly/react-core/dist/esm/components/Wizard/index.js";
 import { Flex, FlexItem } from "@patternfly/react-core/dist/esm/layouts/Flex/index.js";
-import { Stack } from "@patternfly/react-core/dist/esm/layouts/Stack/index.js";
 
 import { getDeviceChildren } from "../../helpers/storage.js";
 
@@ -38,13 +37,16 @@ import { usePageComplete as useDatetimePageComplete } from "../datetime/usePageC
 import { InstallationLanguageReviewDescription } from "../localization/index.js";
 import { usePageComplete as useLocalizationPageComplete } from "../localization/usePageComplete.js";
 import { usePageComplete as useSoftwarePageComplete } from "../software/usePageComplete.js";
+import {
+    StorageInstallationReviewSummary,
+    StorageScenarioReviewDescription,
+} from "../storage/index.js";
 import { useScenario } from "../storage/installation-method/InstallationScenario.jsx";
 import { usePageComplete as useStorageInstallationPageComplete } from "../storage/installation-method/usePageComplete.jsx";
 import { AccountsReviewDescription } from "../users/index.js";
 import { usePageComplete as useUsersPageComplete } from "../users/usePageComplete.jsx";
 import { ReviewDescriptionListItem } from "./Common.jsx";
 import { HostnameRow } from "./Hostname.jsx";
-import { StorageReview, StorageReviewNote } from "./StorageReview.jsx";
 
 import "./ReviewConfiguration.scss";
 
@@ -77,8 +79,6 @@ const ReviewDescriptionList = ({ children }) => {
 
 export const ReviewConfiguration = ({ automatedInstall }) => {
     const osRelease = useContext(OsReleaseContext);
-    const { getLabel } = useScenario();
-    const scenarioLabel = getLabel?.({ isReview: true });
     const userInterfaceConfig = useContext(UserInterfaceContext);
     const hiddenScreens = userInterfaceConfig.hidden_webui_pages || [];
     const isBootIso = useContext(SystemTypeContext).systemType === "BOOT_ISO";
@@ -131,6 +131,8 @@ export const ReviewConfiguration = ({ automatedInstall }) => {
         ? <DateAndTimeReviewDescription />
         : <IncompleteStepIndicator />;
 
+    const installationScenarioDescription = <StorageScenarioReviewDescription />;
+
     const softwareDescription = useMemo(() => {
         if (!softwareSelectionComplete) {
             return <IncompleteStepIndicator />;
@@ -143,15 +145,15 @@ export const ReviewConfiguration = ({ automatedInstall }) => {
         return env?.name || envId;
     }, [softwareSelectionComplete, environments, selection?.environment]);
 
-    const installationTypeDescription = scenarioLabel;
     const storageDescription = (
-        <>
-            <Stack hasGutter>
-                <StorageReview isReviewScreen />
-                {!storageComplete && <IncompleteStepIndicator />}
-            </Stack>
-            <StorageReviewNote />
-        </>
+        <Flex direction={{ default: "column" }} spaceItems={{ default: "spaceItemsSm" }}>
+            {!storageComplete && (
+                <FlexItem>
+                    <IncompleteStepIndicator />
+                </FlexItem>
+            )}
+            <StorageInstallationReviewSummary />
+        </Flex>
     );
 
     const accountDescription = usersComplete
@@ -244,7 +246,7 @@ export const ReviewConfiguration = ({ automatedInstall }) => {
                                     <ReviewDescriptionListItem
                                       id={`${SCREEN_ID}-target-system-mode`}
                                       term={_("Installation type")}
-                                      description={installationTypeDescription}
+                                      description={installationScenarioDescription}
                                     />
                                 </ReviewDescriptionList>
                                 <ReviewDescriptionList>
