+++ src/components/storage/index.js
@@ -1,0 +1,9 @@
+/*
+ * Copyright (C) 2026 Red Hat, Inc.
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ */
+
+import { StorageInstallationReviewSummary } from "./installation-method/StorageInstallationReviewSummary.jsx";
+import { StorageScenarioReviewDescription } from "./installation-method/StorageScenarioReviewDescription.jsx";
+
+export { StorageInstallationReviewSummary, StorageScenarioReviewDescription };
